enable sleep, upgrade to Arche v0.6.2
