+++ go.mod
@@ -3,7 +3,7 @@ module github.com/mlange-42/arche-model
 go 1.20
 
 require (
-	github.com/mlange-42/arche v0.6.1
+	github.com/mlange-42/arche v0.6.2
 	github.com/stretchr/testify v1.8.1
 	golang.org/x/exp v0.0.0-20230213192124-5e25df0256eb
 )
